pyproject.toml
[project]
name = "eva"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "google-search-results>=2.4.2",
    "hatchling>=1.27.0",
    "langchain>=0.3.24",
    "langchain-community>=0.3.23",
    "langchain-openai>=0.3.14",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
    "typer>=0.15.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
    "src/eva",
    "src/eva/commands",
    "src/eva/core",
    "src/eva/vulns",
    ]

[project.scripts]
eva = "eva:app"
